requirements.txt
fastapi==0.110.0
pydantic==2.6.1
uvicorn==0.27.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pytest==8.0.0
pytest-cov==4.1.0
httpx>=0.24.0
colorama==0.4.6
Flask==2.3.3
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
bcrypt==4.0.1
python-multipart==0.0.9
openai==1.12.0
pillow==10.2.0
email-validator==2.1.0
pydantic-settings==2.1.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
h11==0.14.0
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
pydantic_core==2.16.2
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.10.0
pylint==3.0.3
black==24.1.1
isort==5.13.2
mypy==1.8.0
psutil==5.9.8
requests==2.32.3
PyJWT==2.8.0
pytest-asyncio==0.23.5 